+++ src/main/java/wooteco/subway/ui/ExceptionAdvice.java
@@ -5,8 +5,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import wooteco.subway.dto.ErrorResponse;
 import wooteco.subway.exception.DataNotExistException;
+import wooteco.subway.ui.response.ErrorResponse;
 
 @RestControllerAdvice
 public class ExceptionAdvice {
