+++ src/main/java/wooteco/subway/ui/request/LineRequest.java
@@ -1,6 +1,7 @@
-package wooteco.subway.dto;
+package wooteco.subway.ui.request;
 
 public class LineRequest {
+
     private String name;
     private String color;
     private Long upStationId;
