+++ src/main/java/wooteco/subway/ui/request/SectionRequest.java
@@ -1,6 +1,7 @@
-package wooteco.subway.dto;
+package wooteco.subway.ui.request;
 
 public class SectionRequest {
+
     private Long upStationId;
     private Long downStationId;
     private int distance;
