+++ src/test/java/wooteco/subway/acceptance/LineAcceptanceTest.java
@@ -17,7 +17,7 @@ import org.springframework.http.MediaType;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import wooteco.subway.dto.LineResponse;
+import wooteco.subway.ui.response.LineResponse;
 
 @DisplayName("지하철 노선 관련 기능")
 public class LineAcceptanceTest extends AcceptanceTest {
