+++ src/test/java/wooteco/subway/acceptance/StationAcceptanceTest.java
@@ -16,7 +16,7 @@ import org.springframework.http.MediaType;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import wooteco.subway.dto.StationResponse;
+import wooteco.subway.ui.response.StationResponse;
 
 @DisplayName("지하철역 관련 기능")
 public class StationAcceptanceTest extends AcceptanceTest {
